+++ bin/cmds/babel.js
@@ -1,0 +1,3 @@
+exports.command = 'babel <command>';
+exports.desc = 'Manage babel configuration';
+exports.builder = (yargs) => yargs.commandDir('babel_cmds');