+++ bin/cmds/documentation.js
@@ -1,0 +1,4 @@
+exports.command = ['documentation <command>', 'doc'];
+exports.desc = 'Initialize documentation.';
+exports.builder = (yargs) => yargs.commandDir('documentation_cmds');
+