+++ bin/cmds/lint.js
@@ -1,0 +1,3 @@
+exports.command = 'lint <command>';
+exports.desc = 'Manage linting configuration';
+exports.builder = (yargs) => yargs.commandDir('lint_cmds');