+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 ## [?NEXT?] - ????-??-??
 
+- Added rollup-umd bin and dependencies.
+- Documentation variables build in pages job.
+- Added doc prerequisite and commands.
+- Move from group `styled-components` to `bootstrap-styled`.
+- Fix `BootstrapProvider` not updating theme.
+
+## [v1.6.0] - 2018-04-16
+
 - Add dependency [ci-publish](https://www.npmjs.com/package/ci-publish) ^1.3.1.
 - Add dependency [basic-math-utils](https://www.npmjs.com/package/basic-math-utils) ^0.3.4.
 - Remove dependency [math-utils](https://module.kopaxgroup.com/dev-tools/math-utils/tags/v0.3.1).